Fix: Alteração do Valor da Diaria de BigDecimal para Double, correção de nullPoint ao excluir predio e buscar todas localidades. Validação na localidade controller

+++ ms-quartos/src/main/java/br/com/fiap/hackathon/quartos/dtos/QuartoDto.java
@@ -33,7 +33,7 @@ public class QuartoDto {
       value = "0.0",
       inclusive = false,
       message = "O valor da diária deve ser maior que zero")
-  private BigDecimal valorDiaria;
+  private Double valorDiaria;
 
   @Schema(description = "Lista de móveis disponíveis no quarto", required = true)
   @NotEmpty(message = "A lista de móveis não pode estar vazia")
